backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1

# Text extraction
pymupdf==1.24.1
python-docx==1.1.0

# AI / Embeddings — using Google Gemini API (no local model needed)
faiss-cpu>=1.7.4
google-generativeai==0.5.4

# Database — PostgreSQL (production) / SQLite (local dev)
sqlalchemy==2.0.29
asyncpg>=0.29.0          # PostgreSQL async driver (production)
aiosqlite==0.20.0        # SQLite async driver (local dev fallback)


# Utilities
pydantic==2.7.0
numpy<2.0
